Patch Horizontal Server List
Just a quick and simple patch for the theme HorizontalServerList. Since a while, when hovered, the server list extends only across half of the screen. I don't know if that's because another plugin or something else, but this simple change fixes it. Also, it shouldn't interfere with anything else since I used the theme detection already present.